アラートを設ける
