+++ app/javascript/components/Team/Table.js
@@ -4,6 +4,8 @@ import axios from 'axios';
 import EditButton from '../Button/Edit';
 import TrashButton from '../Button/Trash';
 import Form from './Form';
+import SuccessAlert from '../Alert/Sucess';
+import DangerAlert from '../Alert/Danger';
 
 const Table = ({ teams }) => {
   const [name, setName] = useState('');
@@ -18,17 +20,17 @@ const Table = ({ teams }) => {
         if (response.status === 200) {
           setData([...data, response.data.team]);
           setName('');
-          const alert = document.getElementById('alert');
+          const alert = document.getElementById('success-alert');
           alert.classList.remove('hidden');
           setAlert({ status: response.status, message: response.data.message })
         } else {
-          const alert = document.getElementById('alert');
+          const alert = document.getElementById('danger-alert');
           alert.classList.remove('hidden');
           setAlert({ status: response.status, message: response.data.message })
         }
       });
     } else {
-      const alert = document.getElementById('alert');
+      const alert = document.getElementById('danger-alert');
       alert.classList.remove('hidden');
       setAlert(Object.assign({}, { status: 400, message: 'チーム名を入力してください' }))
     }
@@ -41,20 +43,20 @@ const Table = ({ teams }) => {
         .then((response) => {
           if (response.status === 200) {
             const newData = data.map((team) => team.id === response.data.team.id ? response.data.team : team);
-            const alert = document.getElementById('alert');
+            const alert = document.getElementById('success-alert');
             const editForm = document.getElementById(`edit-team-${response.data.team.id}`);
             setData(newData);
             alert.classList.remove('hidden');
             setAlert({ status: response.status, message: response.data.message });
             editForm.classList.add('hidden');
           } else {
-            const alert = document.getElementById('alert');
+            const alert = document.getElementById('danger-alert');
             alert.classList.remove('hidden');
             setAlert({ status: response.status, message: response.data.message });
           }
         });
     } else {
-      const alert = document.getElementById('alert');
+      const alert = document.getElementById('danger-alert');
       alert.classList.remove('hidden');
       setAlert(Object.assign({}, { status: 400, message: 'チーム名を入力してください' }));
     }
@@ -68,21 +70,17 @@ const Table = ({ teams }) => {
           if (response.status === 200) {
             const newData = data.filter(team => team.id !== id);
             setData(newData);
+            const alert = document.getElementById('success-alert');
+            alert.classList.remove('hidden');
+            setAlert({ status: response.status, message: response.data.message });
           } else {
-            console.log(response.status)
-            console.log(response.message)
-            const alert = document.getElementById('alert');
+            const alert = document.getElementById('danger-alert');
             alert.classList.remove('hidden');
             setAlert(Object.assign({}, { status: 400, message: 'チーム名を入力してください' }));
           }
         });
     }
   };
-
-  const handleClickCloseAlert = () => {
-    const alert = document.getElementById('alert');
-    alert.classList.add('hidden');
-  }
 
   const handleClickToggleForm = (id) => {
     const editForm = document.getElementById(`edit-team-${id}`);
@@ -91,15 +89,8 @@ const Table = ({ teams }) => {
 
   return (
     <>
-      <div className="my-2 bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative hidden" role="alert" id='alert'>
-        <span className="block sm:inline">{alert.message}</span>
-        <button className="absolute top-0 bottom-0 right-0 px-4 py-3 focus:outline-none" onClick={() => handleClickCloseAlert()}>
-          <svg className="fill-current h-6 w-6 text-red-500" role="button" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-            <title>Close</title>
-            <path d="M14.348 14.849a1.2 1.2 0 0 1-1.697 0L10 11.819l-2.651 3.029a1.2 1.2 0 1 1-1.697-1.697l2.758-3.15-2.759-3.152a1.2 1.2 0 1 1 1.697-1.697L10 8.183l2.651-3.031a1.2 1.2 0 1 1 1.697 1.697l-2.758 3.152 2.758 3.15a1.2 1.2 0 0 1 0 1.698z"/>
-          </svg>
-        </button>
-      </div>
+      <SuccessAlert message={alert.message} />
+      <DangerAlert message={alert.message} />
       <article className='mt-2'>
         <label className='block w-full text-gray-700 text-xs font-bold'>
           チーム登録
